+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
++ Added sos/changelog language and rules
 
 ## [2.3.0] - 2026-04-16
 ### Added
@@ -102,6 +104,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Disable duplicit rules
 - Disable no-redeclare rule
 - Allow class member functions without explicit type definition
+
 ### Added
 - Stricter typescript enum checking